+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Python components can take `targetSetup` as a derivation (preferably made with `base.mkTargetSetup`) as well as a set of overrides.
+- `shellCommands` for python components works like the others (can take a set with command_name = bash_code;).
+
 ## [6.0.0] - 2022-04-29
 
 ### Added